Move StreamMapping out of config folder
