+++ src/beamlime/kafka/__init__.py
@@ -1,9 +1,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
 
-from .message_adapter import InputStreamKey, KafkaTopic
+from .stream_mapping import InputStreamKey, KafkaTopic, StreamMapping
 
 __all__ = [
     'InputStreamKey',
     'KafkaTopic',
+    'StreamMapping',
 ]
